Minor changes to how filenames are constructed.
The old way to create filenames caused issues if the test case is stored
in a sub-directory of the directory where the test case is run. This
caused annoying issues with pytest. The new way to create filenames replaces
the "/" character with "_" creating valid filenames in the current working
directory instead.